updated project to use esmodules and updated eslint to respect the .js extension for esmodule type in node
