--- middlewares/index.js
@@ -1,5 +1,0 @@
-const { rateLimiterUsingThirdParty } = require('./rateLimiter');
-
-module.exports = {
-  rateLimiterUsingThirdParty,
-};
