+++ package.json
@@ -14,6 +14,7 @@
   "version": "1.0.0",
   "description": "API to access the 99 names of Allah",
   "main": "index.js",
+  "type": "module",
   "devDependencies": {
     "@babel/helper-compilation-targets": "^7.16.7",
     "eslint": "^8.10.0",
