+++ src/middlewares/index.js
@@ -1,0 +1,3 @@
+import rateLimiterUsingThirdParty from './rateLimiter';
+
+export default rateLimiterUsingThirdParty;
